src/codebase/ingestor/requirements.txt
numpy==1.21.2
pandas==1.3.2
pyarrow==5.0.0
yfinance==0.1.63
requests==2.26.0
lxml==4.6.3
snowflake-connector-python[pandas]==2.6.0
snowflake-sqlalchemy==1.3.1
